Tách Repo

+++ include/repositories/admin/AdminMovieRepository.h
@@ -10,31 +10,31 @@ using namespace std;
 
 /**
  * @brief Repository để xử lý file movies.txt cho Admin Panel
+ * 
+ * Note: Đây là tầng data-access (repositories). UI panel chỉ nên gọi service,
+ * và service gọi repository này.
  */
 class AdminMovieRepository {
 private:
     string filePath;
     vector<vector<string>> data;
-    
+
     vector<string> splitString(const string& str, char delimiter);
     string generateNewId();
-    
+
 public:
     AdminMovieRepository(const string& path);
-    
-    // Load/Save
+
     void loadFromFile();
     void saveToFile();
-    
-    // CRUD operations
+
     void addRecord(const vector<string>& record);
     void updateRecord(int index, const vector<string>& record);
     void deleteRecord(int index);
-    
-    // Getters
+
     vector<vector<string>> getAllData() const { return data; }
     vector<string> getRecord(int index) const;
-    int getRecordCount() const { return data.size(); }
+    int getRecordCount() const { return (int)data.size(); }
 };
 
 #endif
